+++ Project_2_regression_code.py
@@ -62,346 +62,6 @@ print(f'Number of observations: {N}')
 C = 7
 classNames = ['building_windows_float_processed', 'building_windows_non_float_processed', 'vehicle_windows_float_processed', 'vehicle_windows_non_float_processed', 'containers', 'tableware', 'headlamps']
 
-#%%
-# Crossvalidation (Pretty much copied from excercise 8.1.1)
-
-K = 5
-CV = model_selection.KFold(K, shuffle=True)
-
-
-
-# Values of lambda
-lambdas = np.power(10.0, range(-8, 6))
-
-# Initialize variables
-# T = len(lambdas)
-Error_train = np.empty((K, 1))
-Error_test = np.empty((K, 1))
-Error_train_rlr = np.empty((K, 1))
-Error_test_rlr = np.empty((K, 1))
-Error_train_nofeatures = np.empty((K, 1))
-Error_test_nofeatures = np.empty((K, 1))
-w_rlr = np.empty((M, K))
-mu = np.empty((K, M - 1))
-sigma = np.empty((K, M - 1))
-w_noreg = np.empty((M, K))
-
-k = 0
-lams = []
-for train_index, test_index in CV.split(X, y):
-    # extract training and test set for current CV fold
-    X_train = X[train_index]
-    y_train = y[train_index]
-    X_test = X[test_index]
-    y_test = y[test_index]
-    internal_cross_validation = 10
-    
-    (
-        opt_val_err,
-        opt_lambda,
-        mean_w_vs_lambda,
-        train_err_vs_lambda,
-        test_err_vs_lambda,
-    ) = rlr_validate(X_train, y_train, lambdas, internal_cross_validation)
-    lams.append(opt_lambda)
-    print(f'Optimal lambda: {opt_lambda}')
-
-    # print(X_train.shape)
-
-    # Standardize
-    mu[k, :] = np.mean(X_train[:, 1:], 0)
-    sigma[k, :] = np.std(X_train[:, 1:], 0)
-
-    X_train[:, 1:] = (X_train[:, 1:] - mu[k, :]) / sigma[k, :]
-    X_test[:, 1:] = (X_test[:, 1:] - mu[k, :]) / sigma[k, :]
-
-    # Compute mean squared error without using the input data at all
-    Error_train_nofeatures[k] = (
-        np.square(y_train - y_train.mean()).sum(axis=0) / y_train.shape[0]
-    )
-    Error_test_nofeatures[k] = (
-        np.square(y_test - y_test.mean()).sum(axis=0) / y_test.shape[0]
-    )
-
-    Xty = X_train.T @ y_train
-    XtX = X_train.T @ X_train
-
-    # Estimate weights for the optimal value of lambda, on entire training set
-    lambdaI = opt_lambda * np.eye(M)
-    lambdaI[0, 0] = 0  # Do no regularize the bias term
-    w_rlr[:, k] = np.linalg.solve(XtX + lambdaI, Xty).squeeze()
-
-    # Compute mean squared error with regularization
-    Error_train_rlr[k] = (
-        np.square(y_train - X_train @ w_rlr[:, k]).sum(axis=0) / y_train.shape[0]
-    )
-    Error_test_rlr[k] = (
-        np.square(y_test - X_test @ w_rlr[:, k]).sum(axis=0) / y_test.shape[0]
-    )
-
-    # Estimate weights for unregularized linear regression, on entire training set
-    w_noreg[:, k] = np.linalg.solve(XtX, Xty).squeeze()
-    # Compute mean squared error without regularization
-    Error_train[k] = (
-        np.square(y_train - X_train @ w_noreg[:, k]).sum(axis=0) / y_train.shape[0]
-    )
-    Error_test[k] = (
-        np.square(y_test - X_test @ w_noreg[:, k]).sum(axis=0) / y_test.shape[0]
-    )
-    # OR ALTERNATIVELY: you can use sklearn.linear_model module for linear regression:
-    # m = lm.LinearRegression().fit(X_train, y_train)
-    # Error_train[k] = np.square(y_train-m.predict(X_train)).sum()/y_train.shape[0]
-    # Error_test[k] = np.square(y_test-m.predict(X_test)).sum()/y_test.shape[0]
-
-    # Display the results for the last cross-validation fold
-    if k == K - 1:
-        plt.figure(k, figsize=(12, 8))
-        plt.subplot(1, 2, 1)
-        plt.semilogx(lambdas, mean_w_vs_lambda.T[:, 1:], ".-")  # Don't plot the bias term
-        plt.xlabel("Regularization factor")
-        plt.ylabel("Mean Coefficient Values")
-        plt.grid()
-        # You can choose to display the legend, but it's omitted for a cleaner
-        # plot, since there are many attributes
-        # legend(attributeNames[1:], loc='best')
-
-        plt.subplot(1, 2, 2)
-        plt.title("Optimal lambda: 1e{0}".format(np.log10(opt_lambda)))
-        plt.loglog(
-            lambdas, train_err_vs_lambda.T, "b.-", lambdas, test_err_vs_lambda.T, "r.-"
-        )
-        plt.xlabel("Regularization factor")
-        plt.ylabel("Squared error (crossvalidation)")
-        plt.legend(["Train error", "Validation error"])
-        plt.grid()
-
-    # To inspect the used indices, use these print statements
-    # print('Cross validation fold {0}/{1}:'.format(k+1,K))
-    # print('Train indices: {0}'.format(train_index))
-    # print('Test indices: {0}\n'.format(test_index))
-
-    k += 1
-
-print(f'Error test: {Error_test}')
-
-#%%
-# ANN part (Shamelessly copied from excercise 8.2.6)
-
-
-# Normalize data
-X_ANN = stats.zscore(X_ANN)
-y_ANN = X_ANN[:, [1]]  # Refractive index
-
-
-print(type(y_ANN.shape))
-print(type(y.shape))
-print(f'X_ANN shape: {X_ANN.shape}')
-print(y_ANN.shape)
-print(y.shape)
-
-
-N, M = X_ANN.shape
-
-#%%
-print(f'y shape: {y.shape}')
-
-#%%
-# Parameters for neural network classifier
-n_hidden_units = 2  # number of hidden units
-n_replicates = 1  # number of networks trained in each k-fold
-max_iter = 10000
-
-# K-fold crossvalidation
-K = 3  # only three folds to speed up this example
-
-def nn_regression(X, y, n_hidden_units, n_replicates, max_iter):
-CV = model_selection.KFold(K, shuffle=True)
-
-# Setup figure for display of learning curves and error rates in fold
-summaries, summaries_axes = plt.subplots(1, 2, figsize=(10, 5))
-# Make a list for storing assigned color of learning curve for up to K=10
-color_list = [
-    "tab:orange",
-    "tab:green",
-    "tab:purple",
-    "tab:brown",
-    "tab:pink",
-    "tab:gray",
-    "tab:olive",
-    "tab:cyan",
-    "tab:red",
-    "tab:blue",
-]
-# Define the model
-model = lambda: torch.nn.Sequential(
-    torch.nn.Linear(M, n_hidden_units),  # M features to n_hidden_units
-    torch.nn.Tanh(),  # 1st transfer function,
-    torch.nn.Linear(n_hidden_units, 1),  # n_hidden_units to 1 output neuron
-    # no final tranfer function, i.e. "linear output"
-)
-loss_fn = torch.nn.MSELoss()  # notice how this is now a mean-squared-error loss
-print("Training model of type:\n\n{}\n".format(str(model())))
-errors = []  # make a list for storing generalizaition error in each loop
-for k, (train_index, test_index) in enumerate(CV.split(X_ANN, y)):
-    print("\nCrossvalidation fold: {0}/{1}".format(k + 1, K))
-        
-    # Extract training and test set for current CV fold, convert to tensors
-    X_train = torch.Tensor(X_ANN[train_index, :])
-    y_train = torch.Tensor(y_ANN[train_index])
-    X_test = torch.Tensor(X_ANN[test_index, :])
-    y_test = torch.Tensor(y_ANN[test_index])
-
-
-
-    # Train the net on training data
-    net, final_loss, learning_curve = train_neural_net(
-        model,
-        loss_fn,
-        X=X_train,
-        y=y_train,
-        n_replicates=n_replicates,
-        max_iter=max_iter,
-    )
-
-    print("\n\tBest loss: {}\n".format(final_loss))
-
-    # Determine estimated class labels for test set
-    y_test_est = net(X_test)
-    print(y_test_est.shape)
-
-    # Determine errors and errors
-    se = (y_test_est.float() - y_test.float()) ** 2  # squared error
-    print(f'Se shape: {se.shape}')
-    mse = (sum(se).type(torch.float) / len(y_test)).data.numpy()  # mean
-    print(f'MSE shape: {mse.shape}')
-    print(f'y_test shape: {y_test.shape}')
-    errors.append(mse)  # store error rate for current CV fold
-
-    # Display the learning curve for the best net in the current fold
-    (h,) = summaries_axes[0].plot(learning_curve, color=color_list[k])
-    h.set_label("CV fold {0}".format(k + 1))
-    summaries_axes[0].set_xlabel("Iterations")
-    summaries_axes[0].set_xlim((0, max_iter))
-    summaries_axes[0].set_ylabel("Loss")
-    summaries_axes[0].set_title("Learning curves")
-
-
-# #%%
-# # Correct the length of the errors to be the same
-# min_length = min(len(e) for e in errors)
-# errors = [e[:min_length] for e in errors]
-
-# print(X_ANN.shape)
-# #%%
-# for i in errors:
-#     print(i.shape)
-
-
-# Display the MSE across folds
-summaries_axes[1].bar(
-    np.arange(1, K + 1), np.squeeze(np.asarray(errors)), color=color_list
-)
-summaries_axes[1].set_xlabel("Fold")
-summaries_axes[1].set_xticks(np.arange(1, K-1 + 1))
-summaries_axes[1].set_ylabel("MSE")
-summaries_axes[1].set_title("Test mean-squared-error")
-print("Diagram of best neural net in last fold:")
-weights = [net[i].weight.data.numpy().T for i in [0, 2]]
-biases = [net[i].bias.data.numpy() for i in [0, 2]]
-tf = [str(net[i]) for i in [1, 2]]
-
-
-
-#%%
-draw_neural_net(weights, biases, tf, attribute_names=attributeNames)
-#%%
-# Print the average classification error rate
-print(
-    "\nEstimated generalization error, RMSE: {0}".format(
-        round(np.sqrt(np.mean(errors)), 4)
-    )
-)
-
-# When dealing with regression outputs, a simple way of looking at the quality
-# of predictions visually is by plotting the estimated value as a function of
-# the true/known value - these values should all be along a straight line "y=x",
-# and if the points are above the line, the model overestimates, whereas if the
-# points are below the y=x line, then the model underestimates the value
-plt.figure(figsize=(10, 10))
-y_est = y_test_est.data.numpy()
-y_true = y_test.data.numpy()
-axis_range = [np.min([y_est, y_true]) - 1, np.max([y_est, y_true]) + 1]
-plt.plot(axis_range, axis_range, "k--")
-plt.plot(y_true, y_est, "ob", alpha=0.25)
-plt.legend(["Perfect estimation", "Model estimations"])
-plt.title("Alcohol content: estimated versus true value (for last CV-fold)")
-plt.ylim(axis_range)
-plt.xlim(axis_range)
-plt.xlabel("True value")
-plt.ylabel("Estimated value")
-plt.grid()
-
-plt.show()
-
-print("Ran Exercise 8.2.5")
-
-
-
-#%%
-# Display results and compare models
-print("Linear regression without feature selection:")
-print("- Training error: {0}".format(Error_train.mean()))
-print("- Test error:     {0}".format(Error_test.mean()))
-print("- Test error min:     {0}".format(np.min(Error_test)))
-print("Optimal lambda: {0}".format(lams.index(np.min(lams))))
-print(
-    "- R^2 train:     {0}".format(
-        (Error_train_nofeatures.sum() - Error_train.sum())
-        / Error_train_nofeatures.sum()
-    )
-)
-print(
-    "- R^2 test:     {0}\n".format(
-        (Error_test_nofeatures.sum() - Error_test.sum()) / Error_test_nofeatures.sum()
-    )
-)
-print("Regularized linear regression:")
-print("- Training error: {0}".format(Error_train_rlr.mean()))
-print("- Test error:     {0}".format(Error_test_rlr.mean()))
-print("- Test error min:     {0}".format(np.min(Error_test_rlr)))
-print(
-    "- R^2 train:     {0}".format(
-        (Error_train_nofeatures.sum() - Error_train_rlr.sum())
-        / Error_train_nofeatures.sum()
-    )
-)
-print(
-    "- R^2 test:     {0}\n".format(
-        (Error_test_nofeatures.sum() - Error_test_rlr.sum())
-        / Error_test_nofeatures.sum()
-    )
-)
-
-print(f'\n Regularization improvement: {(np.mean(Error_test) - np.mean(Error_test_rlr)) / np.mean(Error_test) * 100} % \n')
-
-print("Baseline model (mean):")
-print(
-    "- Training error: {0}".format(Error_train_nofeatures.mean())
-)
-print(
-    "- Test error:     {0}".format(Error_test_nofeatures.mean())
-)
-
-#%%
-print("Weights in last fold:")
-
-weightNames = np.concatenate(
-    (["bias"], attributeNames[:-1]), 0
-)  # Add bias term to the names
-for m in range(M):
-    print("{:>15} {:>15}".format(weightNames[m], np.round   (np.mean(w_rlr[m, :]), 4)))
-
-
